Allow multiple concurrent race instances by fixing event registration conflict. Add handling for forfeit/disconnect
There was an issue which prevented creating more than one Race instance. After starting one Race instance another one could not be created. This is now fixed. Also added handling for when a player forfeits or disconnects. Based on if they were leading or not elo calculations are done accordingly.