Add toggle for using custom finish line
Added a toggle for using the track's built in finish line, or a custom finish line. When choosing a custom finish line, this line now acts the same as the track's finish line as far as the plugin is concerned. Custom finish line is still hard coded in timing.lua at the moment.